FABGW-21: Return block number in CommitStatusResponse
Refactor Gateway's CommitStatus service in line with updated protobufs, so the returned CommitStatusResponse message includes the block number in which the transaction committed. This required a refactor to the commit finder implementation and a switch from Counterfeiter to Mockery as the mocking framework for that package due to cyclic import limitations with Counterfeiter.

Signed-off-by: Mark S. Lewis <mark_lewis@uk.ibm.com>